refactor(chat): extract MessageTreeDelegate from ChatViewModel
Move branch navigation (switchBranch), the in-place streaming anchor
(anchorStreamTo), the temporary-chat toggle, and temp-chat in-memory
finalization into a dedicated MessageTreeDelegate, following the existing
ChatStateHandle delegate pattern. Public switchBranch/toggleTemporaryChat
stay as thin wrappers so screen method references are unchanged.

Pure state transforms over ChatStateHandle; no behavior change. Also adds
MODULARIZATION.md cataloguing the decomposition backlog.
